move the hooksvc to a more logical location
